IMPLEMENTACIÓN EN CLASE

+++ proyecto tienda/CLASES/clproducto.cs
@@ -9,7 +9,7 @@ namespace proyecto_tienda.CLASES
     public class clproducto
     {
         public int PRO_ID { get; set; }
-        public float PRO_PRECIO { get; set; }
+        public double PRO_PRECIO { get; set; }
         public int PRO_UNIDAD { get; set; }
         public int PRO_COM_ID { get; set; }
         public int PRO_GRU_ID { get; set; }
